refactor: split agent graph into package
